Change Vim initialization and mappings
Refactor the Vim initialization script to conditionally load a plugin
based on the existence of the 'g:vscode' variable. This allows the code
to be compatible with Visual Studio Code as well. Additionally, set the
leader key to space and configure various Vim sanity options.

The 'dotfiles/nvim/nvim' file is no longer needed, so it was removed.

Finally, renamed the 'lib/99-vim-mode.zsh' file to 'lib/95-vim-mode.zsh'
for consistency with other files in the 'lib' directory.